* nasl/exec.c (cell2int3, cell2int, cell2intW, int2cell): Change returned variable from int to long int. (cell_cmp, nasl_exec): Use long int instead of int for nasl integer variables.
* nasl/nasl_grammar.y (YYSTYPE): Change num type to long int.

* nasl/nasl_tree.h (tree_cell): Change i_val type to long int.

* nasl/nasl_tree.c (dump_cell_val): Adjust for i_val type change.

* nasl/nasl_var.h (anon_nasl_var): Change v_int type to long int.

* nasl/nasl_var.c (array2str, var2str): Adjust for v_int type change.

git-svn-id: svn+ssh://scm.wald.intevation.org/openvas/trunk/openvas-libraries@26558 423fd1db-d629-0410-8442-d21db03e70f4

+++ nasl/nasl_tree.h
@@ -111,7 +111,7 @@ typedef struct TC
   union
   {
     char *str_val;
-    int i_val;
+    long int i_val;
     void *ref_val;              /* internal reference */
   } x;
   struct TC *link[4];
